+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 
  ## **v2.0.0** [UNRELEASED]
 
+**ADDED**  array conversion method
+```
+const cars: Car[] = ...
+const carDtos: CarDto[] =  this.convertionService.convertAll(cars, CarDto);
+```
+
 **ADDED**   debug mode. If metamorphosis is registered in debug mode, all logs are displayed in console.
 ```
 import { MetamorphosisNestModule } from '@fabio.formosa/metamorphosis-nest';
@@ -26,7 +32,7 @@ export class MyApp{
 ```
 where myCustomerLogger.debug must be a function `(msg: string) => void`
 
-**FIXED** todo
+**FIXED** none
 
 **BREAKING CHANGE** now import metamorphosisModule as following:
 ```